Polish startup
These delays allow more reliable startup. Without them, the SCD30 may
need to retry the begin() call, and the display is liable to not show
anything until the second setup() after receiving power (after a RST).

This also keeps the purple boot neopixel longer, and sets a message
when waiting for the intitial measurement.